+++ flutter_example/test/expected_results.dart
@@ -1,0 +1,18 @@
+List<String> getExpectedSeparatorTexts() {
+  return [
+    // simple results are filtered with "simple" trigger word
+    "one",
+    "one two",
+    "one two three",
+    "one 1 filtered",
+    "one two 1 filtered",
+    "one two three 1 filtered",
+  ];
+}
+
+List<String> getExpectedInterceptedTexts() {
+  return [
+    ">>> interceptor simple",
+    ">>filtered>> interceptor filtered",
+  ];
+}