+++ flutter_example/test/translations_with_underline_separator_test.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_example/util/translation_grouped_underline_separator.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+import 'expected_results.dart';
+import 'test_app.dart';
+
+void main() {
+  testWidgets('Translation with dot separator test',
+      (WidgetTester tester) async {
+    List<String> expectedTexts = getExpectedSeparatorTexts();
+
+    List<String> Function(BuildContext context) getTranslationList =
+        (BuildContext context) {
+      return [
+        TranslationGroupedUnderlineSeparator.of(context).one.underline.simple,
+        TranslationGroupedUnderlineSeparator.of(context).one.two.underline.simple,
+        TranslationGroupedUnderlineSeparator.of(context).one.two.three.underline.simple,
+        TranslationGroupedUnderlineSeparator.of(context).one.underline.amount([1]),
+        TranslationGroupedUnderlineSeparator.of(context).one.two.underline.amount([1]),
+        TranslationGroupedUnderlineSeparator.of(context).one.two.three.underline.amount([1]),
+      ];
+    };
+    // Build our app and trigger a frame.
+    await tester.pumpWidget(TestApp(
+      localizationDelegate: TranslationGroupedUnderlineSeparator.delegate,
+      getTranslationList: getTranslationList,
+    ));
+    await tester.pumpAndSettle();
+
+    for (var i = 0; i < expectedTexts.length; i++) {
+      expect(find.text(expectedTexts[i]), findsOneWidget);
+    }
+  });
+}