+++ rosetta/lib/src/stone.dart
@@ -1,3 +1,5 @@
+import 'grouping.dart';
+
 /// An annotation that generates code for translation Jsons
 /// located in the directory described in [Stone.path] parameter.
 class Stone {
@@ -17,10 +19,23 @@ class Stone {
   /// JSON assets from the root bundle.
   final String package;
 
+  /// Optional. Describes how the keys should be split up to create
+  /// sub-keys. The configuration is done using the [Grouping] class.
+  ///
+  /// Keys can be split into multiple sub-keys to create a tree-like
+  /// variable structure in order to group resources together.
+  ///
+  /// No grouping will be applied if [null].
+  ///
+  /// The default [Grouping.separator] String is '.' but
+  /// different ones can be declared.
+  final Grouping grouping;
+
   /// Create an annotation that will generate the Helper and
   /// Delegate classes for the translations located at [path].
   const Stone({
     this.path,
     this.package,
+    this.grouping,
   }) : assert(path != null);
 }